Move AllocationsManager to db/models.py
When we start to add other models then we can do this in a somewhat
standard place and move the old ones there as well (for as long as we
still need them).
